Move ProductProvider to different package and add javadoc
