+++ bc-backend/src/test/java/org/jboss/da/bc/impl/ProjectHiearchyCreatorTest.java
@@ -15,8 +15,8 @@ import org.jboss.da.bc.model.backend.ProjectDetail;
 import org.jboss.da.bc.model.backend.ProjectHiearchy;
 import org.jboss.da.communication.scm.api.SCMConnector;
 import org.jboss.da.model.rest.GAV;
-import org.jboss.da.products.backend.api.Product;
-import org.jboss.da.products.backend.impl.AggregatedProductProvider;
+import org.jboss.da.products.api.Product;
+import org.jboss.da.products.impl.AggregatedProductProvider;
 import org.jboss.da.reports.api.VersionLookupResult;
 import org.jboss.da.reports.backend.api.VersionFinder;
 import org.junit.Before;
