+++ reports-backend/src/main/java/org/jboss/da/products/api/ProductArtifacts.java
@@ -1,4 +1,4 @@
-package org.jboss.da.products.backend.api;
+package org.jboss.da.products.api;
 
 import java.util.Set;
