+++ reports-backend/src/main/java/org/jboss/da/products/impl/AggregatedProductProvider.java
@@ -1,11 +1,11 @@
-package org.jboss.da.products.backend.impl;
+package org.jboss.da.products.impl;
 
 import org.jboss.da.listings.model.ProductSupportStatus;
 import org.jboss.da.model.rest.GA;
-import org.jboss.da.products.backend.api.Artifact;
-import org.jboss.da.products.backend.api.Product;
-import org.jboss.da.products.backend.api.ProductArtifacts;
-import org.jboss.da.products.backend.api.ProductProvider;
+import org.jboss.da.products.api.Artifact;
+import org.jboss.da.products.api.Product;
+import org.jboss.da.products.api.ProductArtifacts;
+import org.jboss.da.products.api.ProductProvider;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Any;
