+++ reports-backend/src/main/java/org/jboss/da/reports/impl/ReportsGeneratorImpl.java
@@ -20,11 +20,11 @@ import org.jboss.da.listings.model.rest.RestProductInput;
 import org.jboss.da.model.rest.GA;
 import org.jboss.da.model.rest.GAV;
 import org.jboss.da.model.rest.VersionComparator;
-import org.jboss.da.products.backend.api.Artifact;
-import org.jboss.da.products.backend.api.Product;
-import static org.jboss.da.products.backend.api.Product.UNKNOWN;
-import org.jboss.da.products.backend.api.ProductArtifacts;
-import org.jboss.da.products.backend.impl.AggregatedProductProvider;
+import org.jboss.da.products.api.Artifact;
+import org.jboss.da.products.api.Product;
+import static org.jboss.da.products.api.Product.UNKNOWN;
+import org.jboss.da.products.api.ProductArtifacts;
+import org.jboss.da.products.impl.AggregatedProductProvider;
 import org.jboss.da.reports.api.AdvancedArtifactReport;
 import org.jboss.da.reports.api.AlignmentReportModule;
 import org.jboss.da.reports.api.ArtifactReport;
