+++ reports-backend/src/test/java/org/jboss/da/reports/impl/ReportsGeneratorImplTest.java
@@ -17,10 +17,10 @@ import org.jboss.da.listings.api.service.ProductVersionService;
 import org.jboss.da.listings.api.service.WhiteArtifactService;
 import org.jboss.da.listings.model.ProductSupportStatus;
 import org.jboss.da.model.rest.GAV;
-import org.jboss.da.products.backend.api.Artifact;
-import org.jboss.da.products.backend.api.Product;
-import org.jboss.da.products.backend.api.ProductArtifacts;
-import org.jboss.da.products.backend.impl.AggregatedProductProvider;
+import org.jboss.da.products.api.Artifact;
+import org.jboss.da.products.api.Product;
+import org.jboss.da.products.api.ProductArtifacts;
+import org.jboss.da.products.impl.AggregatedProductProvider;
 import org.jboss.da.reports.api.ArtifactReport;
 import org.jboss.da.reports.api.VersionLookupResult;
 import org.jboss.da.reports.backend.api.DependencyTreeGenerator;
